refactor: remove circular imports
Removes circular imports in plate which were causing the docs build to
fail.
