+++ packages/common/plate/src/index.ts
@@ -5,6 +5,7 @@
 export * from 'zod';
 export * from './api';
 export * from './DirectoryTemplate';
+export * from './InteractiveDirectoryTemplate';
 export * from './util/templateLiterals';
 export {
   type InputOf,
